+++ deployment/ccip/changeset/jobs.go
@@ -1,8 +1,7 @@
-package ccipdeployment
+package changeset
 
 import (
 	"github.com/smartcontractkit/chainlink/deployment"
-	"github.com/smartcontractkit/chainlink/deployment/ccip/changeset"
 	"github.com/smartcontractkit/chainlink/v2/core/capabilities/ccip/validate"
 	"github.com/smartcontractkit/chainlink/v2/core/services/relay"
 )
@@ -27,8 +26,8 @@ func NewCCIPJobSpecs(nodeIds []string, oc deployment.OffchainClient) (map[string
 		if !node.IsBootstrap {
 			spec, err = validate.NewCCIPSpecToml(validate.SpecArgs{
 				P2PV2Bootstrappers:     nodes.BootstrapLocators(),
-				CapabilityVersion:      changeset.CapabilityVersion,
-				CapabilityLabelledName: changeset.CapabilityLabelledName,
+				CapabilityVersion:      CapabilityVersion,
+				CapabilityLabelledName: CapabilityLabelledName,
 				OCRKeyBundleIDs: map[string]string{
 					// TODO: Validate that that all EVM chains are using the same keybundle.
 					relay.NetworkEVM: node.FirstOCRKeybundle().KeyBundleID,
@@ -40,8 +39,8 @@ func NewCCIPJobSpecs(nodeIds []string, oc deployment.OffchainClient) (map[string
 		} else {
 			spec, err = validate.NewCCIPSpecToml(validate.SpecArgs{
 				P2PV2Bootstrappers:     []string{}, // Intentionally empty for bootstraps.
-				CapabilityVersion:      changeset.CapabilityVersion,
-				CapabilityLabelledName: changeset.CapabilityLabelledName,
+				CapabilityVersion:      CapabilityVersion,
+				CapabilityLabelledName: CapabilityLabelledName,
 				OCRKeyBundleIDs:        map[string]string{},
 				// TODO: validate that all EVM chains are using the same keybundle
 				P2PKeyID:     node.PeerID.String(),
